+++ packages/nextjs/src/components/RichText.tsx
@@ -69,7 +69,7 @@ export const RichText = (props: RichTextProps): JSX.Element => {
     if (hasText && !isEditing) {
       initializeLinks();
     }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [hasText]);
 
   const routeHandler = (ev: MouseEvent) => {
@@ -127,7 +127,13 @@ export const RichText = (props: RichTextProps): JSX.Element => {
     });
   };
 
-  return <ReactRichText ref={richTextRef} editable={editable} {...rest} />;
+  return (
+    <ReactRichText
+      ref={richTextRef}
+      editable={editable}
+      {...rest}
+    />
+  );
 };
 
 RichText.displayName = 'NextRichText';